Database Video/requirements.txt
django>=5.1,<7.0
django-environ>=0.11
django-cors-headers>=4.3
gunicorn>=22.0
psycopg2-binary>=2.9
pillow>=10.0
mediapipe>=0.10
opencv-python-headless>=4.8
